+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to
 [Semantic Versioning].
 
+## [Unreleased]
+
+### Added
+- [#6](https://github.com/MAAP-Project/gedi-subsetter/issues/6): Allow user to
+  specify which BEAMs to subset
+
 ## [0.3.0] - 2022-10-31
 
 ### Fixed